# Use default analyzer loader in Roslyn benchmarks