… src/benchmarks/real-world/Roslyn/CoreClrAnalyzerAssemblyLoader.cs