Related to dotnet/roslyn#67604

Fixing the benchmark implementation to work with recently released Roslyn (4.3.0).

For the moment I deleted the custom analyzer loader. If .NET Native can't use this loader then ideally we could just fix that on the Roslyn side since presumably we need to be able to load analyzers in order to run the compiler on .NET Native.