@kunalspathak 2023-04-19
@RikkiGibson - after this update, we saw some regressions on both x64 and arm64. Do you know why?

<img width="744" alt="image" src="https://user-images.githubusercontent.com/12488060/233223260-4474130c-c994-4cd5-8e78-66ef6f0fb446.png">