@RikkiGibson 2023-04-19
Moving from 3.8 to 4.3 brings in 2 years of additional work on the compiler. I think any major changes that happened in that timeframe could have done it.

Perhaps we should compare perf traces from before/after this PR? @dotnet/roslyn-compiler