@kunalspathak 2023-04-19
Here is another one from https://pvscmdupload.blob.core.windows.net/reports/allTestHistory%2f8.0_x64_Windows%2010.0.18362_RunKind%3droslyn%2fCompilerBenchmarks.StageBenchmarks.GetDiagnosticsWithAnalyzers.html

<img width="638" alt="image" src="https://user-images.githubusercontent.com/12488060/233223524-39662be2-f603-4764-bc89-5102f6432a6d.png">

and from https://pvscmdupload.blob.core.windows.net/reports/allTestHistory%2f8.0_x64_Windows%2010.0.18362_RunKind%3droslyn%2fCompilerBenchmarks.StageBenchmarks.SerializeMetadata.html

<img width="636" alt="image" src="https://user-images.githubusercontent.com/12488060/233223594-6d1bdb49-f653-4197-8805-014c8c5cd321.png">