@RikkiGibson 2023-04-20
Do the perf machines take ETL traces or anything like that for these scenarios? Otherwise, I imagine we will just have to pull down the before/after versions of the performance repo locally, take a trace, and compare?