@kunalspathak 2023-04-20
@LoopedBard3  can you help @RikkiGibson in getting ETL traces, if we are not already doing it?