@kunalspathak 2023-04-20
> @kunalspathak was this auto filed?

I am not sure, since we do not triage real world app's performance issues because of noise level. This was noticed from the newly created real-world benchmarks dashboard by @LoopedBard3 .