@jaredpar 2023-04-20
Is there a guide on how to use the website with the perf data? 

- When I select a range a popup says "double tap to zoom out". I've double tapped everywhere on the screen, nothing has changed. 
- When I click on markers on the graph I can "Set Baseline" and "Set Compare". Again though nothing changes. 
- If there are ETL files to look at they are _really_ hard to find.