@jaredpar 2023-04-20
Also how do we see what command was executed in both cases?