@LoopedBard3 2023-04-20
@RikkiGibson 
We have two other views depending on how in-depth you are looking to go looking into the data. The first is our PowerBI reports at [aka.ms/dotnetperfstatus](aka.ms/dotnetperfstatus). In this case, the Roslyn data is available in the RealWorldScenarios report found on the left side of the page. Once selected, use the Runkind, ShortTestName to select the test you want to view and then the other selectors to filter the exact run setup. Let me know if you are not able to access the PowerBI. 

We also have an Azure Data Explorer Dashboard here: [here](https://dataexplorer.azure.com/dashboards/bdd5e46a-6534-43b2-a280-1a32ead567cc?p-_startTime=60days&p-_endTime=now&p-_repo=v-dotnet%2Fcore-sdk&p-_branch=v-8.0&p-_queue=v-Windows.10.Amd64.19H1.Tiger.Perf&p-_buildArch=v-x64&p-_runKind=v-roslyn&p-_compMode=v-tiered&p-_otherConfigs=v-%7B%7D&p-_test=v-CompilerBenchmarks.RoslynApis.BuildAnalyzerConfigs#66e0f64f-9a80-4599-931d-467c577e2e63). From this page you can further explore the backing data by clicking the three dots and then explore query on the graph. For this you will need to join a security group, let me know if you want to try the ADX dashboard.