@jaredpar 2023-04-20
@LoopedBard3 

The link you posted seems incorrect. The text is correct but the link is wrong. This seems to be the correct link: https://aka.ms/dotnetperfstatus. I've looked at the PowerBI and it's unclear how to get it to show this specific regression. 

What would really help us is the ETL of the before and after. How do we get that?

Also how can we setup the PowerBI so that we can see the before after for `GetDiagnosticsWithAnalyzers` compared to `GetDiagnostics` for the same setup? That will help us understand what may be happening here. 

Further how do we see what actually ran here? What command are we actually measuring?