@kunalspathak 2023-04-20
You can try this command locally to repro.

```cmd
cd <repo>\src\benchmarks\real-world\Roslyn
dotnet build -c Release
dotnet <repo>\artifacts\bin\CompilerBenchmarks\Release\net7.0\CompilerBenchmarks.dll  --filter *BuildAnalyzerConfigs*
```

On my local machine, I did the above, reverted this PR and remeasured and I can repro the regression:

With PR:

|               Method |     Mean |    Error |   StdDev |   Median |      Min |      Max |     Gen0 |     Gen1 |    Gen2 | Allocated |
|--------------------- |---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|--------:|----------:|
| BuildAnalyzerConfigs | 28.27 ms | 0.282 ms | 0.405 ms | 28.31 ms | 27.31 ms | 28.88 ms | 437.5000 | 375.0000 | 31.2500 |   2.65 MB |

Without PR:
|               Method |     Mean |    Error |   StdDev |   Median |      Min |      Max |     Gen0 |     Gen1 | Allocated |
|--------------------- |---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|----------:|
| BuildAnalyzerConfigs | 26.95 ms | 0.262 ms | 0.291 ms | 26.96 ms | 26.48 ms | 27.47 ms | 343.7500 | 312.5000 |   2.14 MB |

With that, you should be able to gather the ETL traces to further investigate. Hope that helps!