@kunalspathak 2023-04-20
You can also change the filter to something else depending on what you want to investigate. You can see the full names of class by using:

```cmd
dotnetr <repo>\artifacts\bin\CompilerBenchmarks\Release\net7.0\CompilerBenchmarks.dll --filter * --list flat
CompilerBenchmarks.RoslynApis.BuildAnalyzerConfigs
CompilerBenchmarks.StageBenchmarks.CompileMethodsAndEmit
CompilerBenchmarks.StageBenchmarks.SerializeMetadata
CompilerBenchmarks.StageBenchmarks.GetDiagnostics
CompilerBenchmarks.StageBenchmarks.GetDiagnosticsWithAnalyzers
```

You do not need PowerBI or anything like that to repro this.