@LoopedBard3 2023-04-20
* Fixed the link, didn't know external links had to be prefixed with https://. 
* We don't get ETL files for every run we do unless there is an autofile issue so we need to run the ETL gathering manually. 
* For comparing the two tests in PowerBI, if you control click the selectors for the test names it will allow you to multi-select, and then you can filter based on the other selectors. 
* The actual test code is in this file: https://github.com/dotnet/performance/blob/main/src/benchmarks/real-world/Roslyn/StageBenchmarks.cs, with the supporting code being in that Roslyn folder.