@LoopedBard3 2023-04-20
Thank you @kunalspathak for getting that local run, let me see what else we may need to get ETL files generated!