@jaredpar 2023-04-20
@kunalspathak your local repro though shows a minimal change in performance, this bug says there is a 25% change.