@kunalspathak 2023-04-20
> @kunalspathak your local repro though shows a minimal change in performance, this bug says there is a 25% change.

True, it could vary depending on what other things were running on my local machine when I measured it, which might not be the case on perf machines.