@kunalspathak 2023-04-20
> @kunalspathak the issue isn't reproducing locally and the official measurements won't provide ETL files. How are we supposed to investigate this?

It doesn't reproduce to the 25% extent, but locally it is 4%. Are you not able to reproduce this locally on your machine?

@LoopedBard3 - If you have access to the perf lab machine, can you see if it reproduces to the similar extent and if yes, share the ETL?