@LoopedBard3 2023-04-21
I have access to the lab machines so I will check to see if I can repro this regression and generate ETL files 👍.